# docs/openapi.yaml
openapi: 3.1.0
info:
  title: api2.autisme.lu
  version: 1.0.0
  description: Modern Node.js API for NamNam and legacy Autisme Luxembourg workflows.
servers:
  - url: https://api2.autisme.lu
security:
  - bearerAuth: []
paths:
...
  /places/{id}:
    get:
      summary: Get place/service
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Place }
    patch:
      summary: Update place/service
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '200': { description: Updated place }
...
  /weekly-menus:
    get:
      summary: Get weekly menu
      parameters:
        - { name: monday_date, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: Weekly menu }
    post:
      summary: Upsert weekly menu
      responses:
        '201': { description: Saved weekly menu }
...
  /weekly-menus/latest/pdf/{type}:
    get:
      summary: Latest PDF export redirect
      parameters:
        - { name: type, in: path, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect }
...
  /reports/generated/{id}:
    delete:
      summary: Delete generated report
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer }
...